+++ CHANGELOG.md
@@ -9,7 +9,7 @@
 * Fixed hand grenade animations not playing correctly [#209](https://github.com/SamVanheer/halflife-updated/pull/209) (Thanks Toodles2You)
 * Fixed out of bounds access in studiomodel renderer bone setup code (halflife issue [#3360](https://github.com/ValveSoftware/halflife/issues/3360))
 * Fixed mouse cursor being invisible in VGUI1 menus when raw input is enabled [#211](https://github.com/SamVanheer/halflife-updated/issues/211) (Thanks RykahKnight)
-* Removed now unnecessary workaround to stop mouse movement on the pause menu from changing in-game angles when unpausing
+* Mouse movement in the main menu no longer affects in-game view angles when not using raw input
 * Fixed RPG being flagged as unusable while a rocket is loaded [#213](https://github.com/SamVanheer/halflife-updated/pull/213) (Thanks Toodles2You)
 * Have clients select weapons by ID, rather than by name [#217](https://github.com/SamVanheer/halflife-updated/pull/217) (Thanks Toodles2You)
 * Copy delta.lst when building client or server to ensure mods have correct delta.lst file (Thanks P38TaKjYzY)